Extract useTimer to useMediaTimer hook
